+++ CHANGELOG.md
@@ -23,6 +23,8 @@ speaking, considered to be the player's experience.
 - Fixed the buttons of the Harbour interface overlapping with the production
   progress bar.
 - Fixed the Harbour production count not being placed properly on the button.
+- Fixed mines built outside of natural resources showing "null" on their
+  information panels.
 
 ### Changed