+++ CHANGELOG.md
@@ -13,6 +13,7 @@ speaking, considered to be the player's experience.
 ### Added
 
 - Automatically detect the language on Windows.
+- Added a Dutch translation, thanks to Jan Miedema.
 - Show the average target resistance of independent towns on the information
   panel when you have workers.