+++ CHANGELOG.md
@@ -10,6 +10,9 @@ speaking, considered to be the player's experience.
 
 ## [Unreleased]
 
+
+## [3.2.7] — 2025-10-24
+
 ### Added
 
 - Automatically detect the language on Windows.